+++ jsonargparse_tests/test_signatures.py
@@ -350,9 +350,8 @@ def test_add_class_group_config(parser, tmp_cwd):
 
 def test_add_class_group_config_not_found(parser, tmp_cwd):
     parser.add_class_arguments(Class0, "a")
-    with pytest.raises(ArgumentError) as ctx:
+    with pytest.raises(ArgumentError, match="Unable to load config .does_not_exist.yaml"):
         parser.parse_args(["--a=does_not_exist.yaml"])
-    ctx.match('Unable to load config "does_not_exist.yaml"')
 
 
 class WithDocstring: